Refactor create context crate to contain the context implementations
For easier navigation the context definitions are moved to a context
crate.
